Fixed XSD location in ErnParserController
When setXsdValidation called the following PHP Warning would be thrown:

`PHP Warning:  XMLReader::setSchema(): I/O warning : failed to load external entity "xsd/release_notification/382/release-notification.xsd" in */../src/Controller/ErnParserController.php on line 683`

This is because the path wasn't valid and needed to drop back a couple of directories, so I changed this to use a relative path, which broke the
tests because the path in the test was wrong as an update. I updated the tests to make 2 assertions, one for the test each side of the dynamic
part.